2018: day 19 part 2
idk this janky. i figured out that the start was just a jump to the 2nd half which did some initialization things, and then the middle of the program was some loop. i tried turning the main loop into pseudocode but wasnt really able to piece together what it was doing. just set some print statements at various points that seemed interesting and figured out what it was doing eventually