2023: day 9
11:55/12:23 | lost an ungodly amount of time because my ints function
did not take negatives. i'm done using it at this point because it's
lost me a ton of time on literally every day i've tried to use it, so
stupid. had an almost working solution for part 1 at ~3m and then spent
the next ~8 minutes trying to figure out why it no worky (there was one
other small bug besides not having negatives, but figuring out that my
parsed input had no negatives was the main bottleneck)

also exactly #999 on part 2 which is funny