2020: day 20
~11m/??h | part 1 was fine, i just took a few minutes to realize the
trick.

i spent several hours trying a weird approach to part 2 where i would
build the grid in concentric rings but it didn't find a solution so i
didn't debug it and gave up. that (non- working) solution is in the
extras folder because it had a lot of comments and i thought it was
actually fairly comprehensible given what it was doing so im proud of
it.

i came back to this problem today after 2 weeks of avoiding it. i
decided to try a new approach of building out 2 sides of the grid and
then filling in the rest of it from there by checking for the missing
common neighbor of each pair of 2 diagonally adjacent tiles until the
grid is full. that code wasn't actually all that bad but after that i
kind of started to lose it and the code just gets worse and worse for
the rest of the problem and it's all bad and i'm not refactoring it this
problem takes way too much thinking i don't want to do it anymore