comp/advent/2024: day 02
4:58 (#492) | 6:45 (#310)

Lost a little over a minute to the server issues (so I imagine this
day's points will be voided anyway), so that probably hurt my placement
a decent amount.

I wasn't expecting day 2 to be this hard and my Python is still a bit
rusty because I haven't done enough practice. The solution in this
commit is cleaned up; I forgot to make use of itertools.pairwise in my
initial solve (and I had some other really roundabout checks for
increasing/decreasing). Overall, the wasted time is mostly my fault, but
the server working would have been nice.