comp/advent/2024: day 05
7:12 (#690) | 12:02 (#541)

Meh, this one was ok. My initial solutions were extremely terrible, but
it's fine. I lost about 4 minutes to a silly oversight (checking against
all requirements for the given number instead of just those in the
update) on part 1, and lost about 2 minutes on part 2 to hackfixing my
code to pass the real input.

Overall, not too bad. There are some cool tricks I still don't make use
of in this cleaned up solution, but I'm ok with it. The problem
description took a lot of time to understand and I didn't have the
energy to think about what properties the update rules must have needed
for the problem to be possible (on 3 hours of sleep).