comp/advent/2024: day 10
2:26 (#116) | 7:59 (#352)

Not a terrible part 1 considering I haven't written a BFS since last
December.

I spent way too long attempting to track the visited list on part 2
which is just... stupid. It works, but it's stupid, and I wasted a ton
of time getting the destructuring and reconstruction wrong before
realizing I could just evaluate all paths within a distance of 9 (the
grid is small and the set of valid paths is incredibly narrow.) Oh well.

Could have made leaderboard if I'd actually practiced and didn't have
brain shutoff in part 2 but meh