comp/advent/2024: day 12 improvement
Not much of an improvement but it's a bit better I think. Should have
thought of this earlier on but I was grappling with other edge cases
before realizing I needed to track which side(s) of the region each
perimeter tile was adjacent to, so this solution would not have been
obvious then.