comp/advent/2024: day 14
4:44 (#160) | 20:31 (#214)

Disappointing, even though it's my 3rd best performance of the year.
I lost about 2 minutes(!!) on part 1 to not reading the fact that bots
on the border of quadrants were not to be counted.

I spent about 9 minutes manually scrolling the output of part 2 before
finally going back to do it more intelligently, since I'd noticed that
there was a repeating pattern every 101 seconds. I then had 2 wrong
submissions to separate off-by-one errors. My method would have still
comfortably landed me in the top 100 for part 2 if I was a bit more
thoughtful, but there were a lot of other clever heuristics people used
which I did not think of (including the one here for "no overlapping
robots", thanks blockog.)