comp/advent/2024: day 15
12:11 (#253) | 39:44 (#319)

This one is certainly one of the problems of the year. A lot harder than
I was expecting for today but meh.

Part 1 was disappointing. I only wrote a single bug (score calculation,
caught immediately) but lost a full minute to comparing against the
wrong sample case and then a full FIVE minutes to not realizing
directions were split over several lines (why??). Without those issues
I would have been in the top 40 or so, which is annoying.

Part 2 went better than I expected given how fiddly this problem is to
get right. I only had to spend about 10 minutes debugging, which was
a relief.

If I hadn't gotten trolled on part 1 this would have been a pretty good
day (top 40 + a near top-100 part 2), but it's still ok-ish.