comp/advent/2024: day 20
35:52 (#1784) | 53:34 (#1162).

Meh, not sure how to feel about this problem. I spent a lot of time
attempting to bruteforce part 1, and it's still not clear to me why it
didn't work, but oh well.

The inputs for this problem are also... seemingly not great? I had
2 fairly major bugs in my initial part 1 solve which didn't cause
any issues, but *did* cause issues on part 2.

I'm also not a fan of how annoying it is to debug this problem and how
much it felt like twiddling with magic constants until it worked.