comp/advent/2024: day 22
2:47 (#135) | 35:22 (#1091)

Small reading errors on part 1 led to a wrong submission, so could have
made top 100 there. I wasted a bit of time attempting to brute-force
part 2 but ended up losing far more time to, once again, not reading the
problem correctly.

That being said, I do think the problem statement for part 2 is pretty
messy relative to all of the other problems so far this year. The
sequences of price differences are pretty terrible, since one of them
ends up placing equal prices on both ends, and the other only has
a difference of 1 before and after the sequence. The initial wording of
when to sell is extremely unhelpful and it's not clear what it means
unless you read the entirety of the following paragraph.