comp/ebc: add everybody codes
Need to update some of my utilities here to make them work with
Everybody Codes, since the format is a bit different from Advent of
Code.

I think it would also probably be a good idea to split out my utility
library into another directory so that it can be shared between Advent
of Code and other programming puzzles, such as Everybody Codes.