comp/ebc/2024: day 05a/05b
8:43 | 11:45 | DNF

This is not a good problem and the inputs (at least on the earlier
parts) are pretty bad, since they allow your solution to be completely
non-working. My solution falls over with inputs whose numbers are larger
than the column sizes, which wasn't an issue on part 1 or 2 but prevents
part 3 and the later test cases from working.