comp/codeforces/2056: problems A, B, C
B is really confusingly worded but I got it eventually. I gave up on
C way too early on and wasted time looking at the other problems, but
figured it out after thinking about it for a few more minutes near the
end of the contest.

Should have solved B and C much faster but it's fine. Happy to solve all
of the first 3 problems in contest, since I didn't manage that on my
first Div 2 round.